create config folder and move database.js into it
Use a config folder permit a better understanding of database.js
